+++ clusterManagement.h
@@ -5,6 +5,6 @@
 #include <stdlib.h>
 #include <math.h>
 int offSetCalc(int8 indexSize,int8 clusterSize,int8 curDir);
-int8 findAvailableCluster(FILE *file, int8 indexSize, int8 clusterSize);
-int8 findNewCluster(FILE *file, int8 indexSize, int8 clusterSize,int8 currentIndex);
+int findAvailableCluster(FILE *file, int8 indexSize, int8 clusterSize);
+int findNewCluster(FILE *file, int8 indexSize, int8 clusterSize,int8 currentIndex);
 #endif
